pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sip-voice-transport"
version = "0.1.0"
description = "PSTN phone call transport for Mac-local voice AI — supports Telnyx and Twilio"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "David Mar"},
]
keywords = ["sip", "voip", "telephony", "voice-ai", "telnyx", "twilio", "webrtc", "transport"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Telephony",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
telnyx = ["telnyx>=2.0"]
twilio = ["twilio>=9.0"]
audio = ["scipy>=1.10", "noisereduce>=3.0"]
cli = ["rich>=13.0", "click>=8.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]
all = ["sip-voice-transport[telnyx,twilio,audio,cli,dev]"]

[project.scripts]
sip-voice-server = "sip_voice_transport.cli.app:main"

[project.urls]
Repository = "https://github.com/davidbmar/sip-voice-transport"

[tool.hatch.build.targets.wheel]
packages = ["src/sip_voice_transport"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
